@@ -12,18 +12,21 @@ class DatasourcesAndWorkbooks(Server):
         super().__init__(args)
 
     @staticmethod
-    def get_view_by_content_url(logger, server, view_content_url):
+    def get_view_by_content_url(logger, server, view_content_url) -> TSC.ViewItem:
+        logger.debug("Fetching view with id {}".format(view_content_url))
         try:
             req_option = TSC.RequestOptions()
             req_option.filter.add(TSC.Filter("contentUrl", TSC.RequestOptions.Operator.Equals, view_content_url))
             matching_views, _ = server.views.get(req_option)
             selected_view = matching_views[0]
-            return selected_view
-        except IndexError:
-            Errors.exit_with_error(logger, "Could not find view. Please check the name and try again.")
+        except Exception as e:
+            Errors.exit_with_error(logger, "Could not find view. Please check the name and try again.", e)
+
+        return selected_view
 
     @staticmethod
-    def get_wb_by_content_url(logger, server, workbook_content_url):
+    def get_wb_by_content_url(logger, server, workbook_content_url) -> TSC.WorkbookItem:
+        logger.debug("fetch workbook with id {}".format(workbook_content_url))
         try:
             req_option = TSC.RequestOptions()
             req_option.filter.add(
@@ -35,6 +38,7 @@ class DatasourcesAndWorkbooks(Server):
             )
             matching_workbooks, _ = server.workbooks.get(req_option)
             selected_workbook = matching_workbooks[0]
-            return selected_workbook
-        except IndexError:
-            Errors.exit_with_error(logger, "Could not find view. Please check the name and try again.")
+        except Exception:
+            Errors.exit_with_error(logger, "Could not find workbook. Please check the name and try again.")
+
+        return selected_workbook
