@@ -3,8 +3,10 @@ import tableauserverclient as TSC
 from tabcmd.commands.auth.session import Session
 from tabcmd.execution.logger_config import log
 from .datasources_and_workbooks_command import DatasourcesAndWorkbooks
+from tabcmd.commands.datasources_and_workbooks.export_command import ExportCommand
 from tabcmd.execution.global_options import *
 from tabcmd.execution.localize import _
+from tabcmd.commands.constants import Errors
 
 
 class GetUrl(DatasourcesAndWorkbooks):
@@ -18,141 +20,175 @@ class GetUrl(DatasourcesAndWorkbooks):
 
     @staticmethod
     def define_args(get_url_parser):
-        get_url_parser.add_argument("url", help=_("get.errors.requires_url"))
+        get_url_parser.add_argument(
+            "url",
+            help="url that identifies the view or workbook to export\n"
+            "e.g: /workbooks/workbook-name.twbx or views/view-name.pdf",
+        )
         set_filename_arg(get_url_parser)
-        # TODO add args for png size in pixels, refresh
-        # to send to the server these are both just set on the url?
+        # these don't need arguments, although that would be a good future addition
+        # tabcmd get "/views/Finance/InvestmentGrowth.png?:size=640,480" -f growth.png
+        # tabcmd get "/views/Finance/InvestmentGrowth.png?:refresh=yes" -f growth.png
 
     @staticmethod
     def run_command(args):
-        logger = log(__class__.__name__, args.logging_level)
-        logger.debug(_("tabcmd.launching"))
+        # A view can be returned in PDF, PNG, or CSV (summary data only) format.
+        # A Tableau workbook is returned as a TWB if it connects to a datasource/live connection,
+        # or a TWBX if it uses an extract.
+        logger = log(__name__, args.logging_level)
+        logger.debug("======================= Launching command =======================")
         session = Session()
         server = session.create_session(args)
-        file_type = GetUrl.evaluate_file_name(logger, args.filename, args.url)
-        if file_type == "pdf":
-            GetUrl.generate_pdf(logger, server, args)
-        elif file_type == "png":
-            GetUrl.generate_png(logger, server, args)
-        elif file_type == "csv":
-            GetUrl.generate_csv(logger, server, args)
-        elif file_type == "twbx" or file_type == "twb":
-            GetUrl.generate_twb(logger, server, args)
-        else:
-            GetUrl.exit_with_error(logger, _("tabcmd.get.extension.not_found"))
-
-    @staticmethod
-    def evaluate_file_name(logger, file_name, url):
-        type_of_file = None
-        if file_name is not None:
-            split_file_name = file_name.split(".")
-            type_of_file = split_file_name[1]
-        else:  # file_name is None:
-            # grab from url
-            split_url_to_get_extension = url.split(".")
-            if len(split_url_to_get_extension) > 1:
-                type_of_file = split_url_to_get_extension[1]
+        file_type = GetUrl.get_file_type_from_filename(logger, args.filename, args.url)
+        content_type = GetUrl.evaluate_content_type(logger, args.url)
+        if content_type == "workbook":
+            if file_type == "twbx" or file_type == "twb":
+                GetUrl.generate_twb(logger, server, args, file_type)
             else:
-                GetUrl.exit_with_error(logger, _("tabcmd.error.extension.not_found"))
-        return type_of_file
+                Errors.exit_with_error(logger, message="A workbook can only be exported as twb or twbx")
+        else:  # content type = view
+            if file_type == "pdf":
+                GetUrl.generate_pdf(logger, server, args)
+            elif file_type == "png":
+                GetUrl.generate_png(logger, server, args)
+            elif file_type == "csv":
+                GetUrl.generate_csv(logger, server, args)
+            else:
+                Errors.exit_with_error(logger, message=_("tabcmd.get.extension.not_found"))
 
     @staticmethod
-    def check_if_extension_present(view_name):
-        split_view = view_name.split(".")
-        if len(split_view) == 2:
-            return split_view[1] in ["pdf", "csv", "png", "twb", "twbx"]
-        return False
-
-    @staticmethod
-    def get_view_without_extension(view_name):
-        split_view = view_name.split(".")
-        return split_view[0]
-
-    @staticmethod
-    def get_workbook(url):
-        separated_list = url.split("/")
-        if GetUrl.check_if_extension_present(separated_list[::-1][0]):
-            view_second_half_url = GetUrl.get_view_without_extension(separated_list[::-1][0])
+    def evaluate_content_type(logger, url):
+        # specify a view to get using "/views/<workbookname>/<viewname>.<extension>"
+        # specify a workbook to get using "/workbooks/<workbookname>.<extension>".
+        if url.find("/views/") == 0:
+            return "view"
+        elif url.find("/workbooks/") == 0:
+            return "workbook"
         else:
-            view_second_half_url = separated_list[2]
-        return view_second_half_url
+            Errors.exit_with_error(
+                logger,
+                message="Content url must be in the format /views/workbook-name/view-name or /workbooks/workbook-name",
+            )
 
     @staticmethod
-    def get_view(url):
-        # check the size of list
-        separated_list = url.split("/")
-        if GetUrl.check_if_extension_present(separated_list[::-1][0]):
-            view_second_half_url = GetUrl.get_view_without_extension(separated_list[::-1][0])
+    def get_file_type_from_filename(logger, file_name, url):
+        type_of_file = None
+
+        if file_name:
+            logger.debug("Get file type from filename: {}".format(file_name))
+            type_of_file = GetUrl.get_file_extension(file_name)
         else:
-            view_second_half_url = separated_list[2]
-        return "{}/sheets/{}".format(separated_list[1], view_second_half_url)
+            logger.debug("Get file type from url: {}".format(url))
+            if url and url.find(".") > 0:
+                type_of_file = GetUrl.get_file_extension(url)
+
+        if not type_of_file:
+            Errors.exit_with_error(logger,  _("tabcmd.error.extension.not_found")")
+
+        logger.debug("extension from command line is {}".format(type_of_file))
+        if type_of_file in ["pdf", "csv", "png", "twb", "twbx"]:
+            logger.debug("Valid file type")
+            return type_of_file
+
+        Errors.exit_with_error(logger, "The file type {} is invalid".format(type_of_file))
+
+    @staticmethod
+    def get_file_extension(filename):
+        parts = filename.split(".")
+        if len(parts) < 2:
+            return None
+        extension = parts[1]
+        extension = GetUrl.strip_query_params(extension)
+        return extension
+
+    @staticmethod
+    def strip_query_params(filename):
+        if filename.find("?") > 0:
+            filename = filename.split("?")[0]
+        return filename
+
+    @staticmethod
+    def strip_extension(filename):
+        if filename.find(".") > 0:
+            filename = filename.split(".")[0]
+        return filename
+
+    @staticmethod
+    def get_workbook_name(logger, url):  # /workbooks/wb-name" -> "wb-name"
+        name_parts = url.split("/")
+        workbook_name = name_parts[::-1][0]  # last part
+        workbook_name = GetUrl.strip_query_params(workbook_name)
+        workbook_name = GetUrl.strip_extension(workbook_name)
+        return workbook_name
+
+    @staticmethod
+    def get_view_url(url):  # "/views/wb-name/view-name" -> wb-name/sheets/view-name
+        name_parts = url.split("/")  # ['', 'views', 'wb-name', 'view-name']
+        if len(name_parts) != 4:
+            raise ValueError("The url given did not match the expected format: 'views/workbook-name/view-name'")
+        view_name = name_parts[::-1][0]
+        view_name = GetUrl.strip_query_params(view_name)
+        view_name = GetUrl.strip_extension(view_name)
+
+        workbook_name = name_parts[2]
+        return "{}/sheets/{}".format(workbook_name, view_name)
+
+    @staticmethod
+    def filename_from_args(file_argument, item_name, filetype):
+        if file_argument is None:
+            file_argument = "{}.{}".format(item_name, filetype)
+        return file_argument
 
     @staticmethod
     def generate_pdf(logger, server, args):
-        view = GetUrl.get_view(args.url)
+        view_url = GetUrl.get_view_url(args.url)
         try:
-            views_from_list = GetUrl.get_view_by_content_url(logger, server, view)
+            view_item: TSC.ViewItem = GetUrl.get_view_by_content_url(logger, server, view_url)
+            logger.debug("Fetched view item {}".format(view_item.name))
             req_option_pdf = TSC.PDFRequestOptions(maxage=1)
-            server.views.populate_pdf(views_from_list, req_option_pdf)
-            if args.filename is None:
-                file_name_with_path = "{}.pdf".format(views_from_list.name)
-            else:
-                file_name_with_path = args.filename
-            formatted_file_name = file_name_with_path
-            with open(formatted_file_name, "wb") as f:
-                f.write(views_from_list.pdf)
-                logger.info(_("export_success"), views_from_list.name, formatted_file_name)
+            server.views.populate_pdf(view_item, req_option_pdf)
+            filename = GetUrl.filename_from_args(args.filename, view_item.name, "pdf")
+            with open(filename, "wb") as f:
+                f.write(view_item.pdf)
+            logger.info(_("export_success"), views_from_list.name, formatted_file_name)
         except TSC.ServerResponseError as e:
-            GetUrl.exit_with_error(logger, _("publish.errors.unexpected_server_response"), e)
+             GetUrl.exit_with_error(logger, _("publish.errors.unexpected_server_response"), e)
 
     @staticmethod
     def generate_png(logger, server, args):
-        view = GetUrl.get_view(args.url)
+        view = GetUrl.get_view_url(args.url)
         try:
-            views_from_list = GetUrl.get_view_by_content_url(logger, server, view)
+            view_item: TSC.ViewItem = GetUrl.get_view_by_content_url(logger, server, view)
             req_option_csv = TSC.CSVRequestOptions(maxage=1)
-            server.views.populate_csv(views_from_list, req_option_csv)
-            if args.filename is None:
-                file_name_with_path = "{}.png".format(view)
-            else:
-                file_name_with_path = args.filename
-            formatted_file_name = file_name_with_path
-            with open(formatted_file_name, "wb") as f:
-                f.write(views_from_list.png)
-                logger.info(_("export_success"), views_from_list.name, formatted_file_name)
+            server.views.populate_csv(view_item, req_option_csv)
+            filename = GetUrl.filename_from_args(args.filename, view_item.name, "png")
+            with open(filename, "wb") as f:
+                f.write(view_item.png)
+            logger.info(_("export_success"), views_from_list.name, formatted_file_name)
         except TSC.ServerResponseError as e:
-            GetUrl.exit_with_error(logger, _("publish.errors.unexpected_server_response"), e)
+             GetUrl.exit_with_error(logger, _("publish.errors.unexpected_server_response"), e)
 
     @staticmethod
     def generate_csv(logger, server, args):
-        view = GetUrl.get_view(args.url)
+        view_url = GetUrl.get_view_url(args.url)
         try:
-            views_from_list = GetUrl.get_view_by_content_url(logger, server, view)
+            view_item: TSC.ViewItem = GetUrl.get_view_by_content_url(logger, server, view_url)
             req_option_csv = TSC.CSVRequestOptions(maxage=1)
-            server.views.populate_csv(views_from_list, req_option_csv)
-            if args.filename is None:
-                file_name_with_path = "{}.csv".format(view)
-            else:
-                file_name_with_path = args.filename
-            formatted_file_name = file_name_with_path
-            with open(formatted_file_name, "wb") as f:
-                f.write(views_from_list.csv)
-                logger.info(_("export_success"), views_from_list.name, formatted_file_name)
+            server.views.populate_csv(view_item, req_option_csv)
+            file_name_with_path = GetUrl.filename_from_args(args.filename, view_item.name, "csv")
+            with open(file_name_with_path, "wb") as f:
+                f.write(view_item.csv)
+            logger.info(_("export_success"), views_from_list.name, formatted_file_name)
         except TSC.ServerResponseError as e:
             GetUrl.exit_with_error(logger, _("publish.errors.unexpected_server_response"), e)
 
     @staticmethod
-    def generate_twb(logger, server, args):
-        workbook = GetUrl.get_workbook(args.url)
+    def generate_twb(logger, server, args, file_extension):
+        workbook_name = GetUrl.get_workbook_name(logger, args.url)
         try:
-            target_workbook = GetUrl.get_view_by_content_url(logger, server, workbook)
-            file_target = None
-            if args.filename is None:
-                file_target = target_workbook.name
-            else:
-                file_target = args.filename
-
-            server.workbooks.download(target_workbook.id, filepath=file_target, no_extract=False)
+            target_workbook = GetUrl.get_wb_by_content_url(logger, server, workbook_name)
+            file_name_with_path = GetUrl.filename_from_args(args.filename, workbook_name, file_extension)
+            server.workbooks.download(target_workbook.id, filepath=file_name_with_path, no_extract=False)
             logger.info(_("export_success"), target_workbook.name, file_target)
         except TSC.ServerResponseError as e:
             GetUrl.exit_with_error(logger, _("publish.errors.unexpected_server_response"), e)
